+++ src/util/data.js
@@ -47,3 +47,7 @@ export const validPromoCode = [
 
 export const taxPercent = 5;
 export const deliveryCost = 20;
+
+export const itemsPerPageCart = 3;
+export const itemsPerPageProduct = 8;
+export const itemsPerPageWishlist = 10;
